up to 4.21 complete

+++ part4/bloglist/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-// helper function to isolate the token from authorization header
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
 
 blogsRouter.get('/', async(request, response) => {
     const blogs = await Blog
@@ -30,13 +22,16 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const token = getTokenFrom(request)
-
+    //console.log('request.body', request.body)
+    //console.log('request.token:', request.token)
+    //const token = getTokenFrom(request)
+    const token = request.token
     // check the validity of the token
     // this method also decodes the token, or returns the Object which the token was based on
     // if there is no token, or the object decoded from th token does not contain user identity, 401
     const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+    //console.log('decodedToken:', decodedToken)
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
@@ -56,8 +51,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }
+    else response.status(401).json({ error: 'user id different from blog creator id' })
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
